msm: scm-boot: Clarify types for scm_set_boot_addr()
The boot address is a physical address (even if the secure
monitor only considers it to be 32 bits). Move to using
phys_addr_t to clarify this is a physical address and to ease
any support of LPAE in the future. Also make the flags unsigned
so that we can use the highest bit if necessary.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
(cherry picked from commit efadd5b002a819d0e0213ac10a1e3d1c9463a07b)

Change-Id: I06c5af4093b6fa947a8705364425de2c54464cde
Signed-off-by: Sudhir Sharma <sudsha@codeaurora.org>