msm: kgsl: Change VBIF settings for adreno A330
Limit the number of pending requests from clients to just 1 for
better stability. More agressive values can be set later if things
are found to be stable.

Switch off gating next write request for certain clients, this
setting seems to be stable and should be better for performance

Set the register values for interleaving requests for A330 VBIF.

Do not set VBIF_OUT_AXI_AMEMTYPE_CONF0 since it was
only being set to POR value.

Change-Id: I2152a30b49545fdbf0c147106a088b71ed562e3f
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Harsh Vardhan Dwivedi <hdwivedi@codeaurora.org>
Signed-off-by: Rajeev Kulkarni <krajeev@codeaurora.org>