msm: clock-debug: Update debug_suspend feature for prepared state
If a clock is prepared and not enabled we don't print information
about that clock in the debug_suspend feature. Fix the code to
consider unprepared as the only case to bail out early and also
update the printk to display the prepare count and enable count.

Change-Id: Ie7b5527c17339e48b93dd6354fed132679351d18
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
(cherry picked from commit 9d96ae5c2bd779f0590565d137030eb4c367a28f)

Signed-off-by: Sudhir Sharma <sudsha@codeaurora.org>