power: pm8921-bms: fix unintended diff between reported SoC and calculated SoC
There is adjustment to prevent discontinuous SoC change when
calculated SoC is reported. This adjustment is performed while
device is charged only. In current implementation, the
determination of charging is not accurate. So in some case,
the adjustment is performed even device is not charged. This cause
unintended difference between reported SoC and calculated Soc.
This patch change the method of charging determination for
reported Soc adjustment.

Change-Id: I7caedf7d696be57c96d5f777e44f8ddecbdbc7f2