Mako: NFC: Return proper error code from address change ioctl.
We're tracking a problem where NFC on Mako does not come up
sometimes. As a software workaround, we want to retry the
reset procedure a few times. Since the user-space stack
calls the address change ioctl first on Mako (which
does some i2c writes), return error code from that
ioctl so we can see if NFC didn't come up.

Bug: 7258919
Change-Id: I01584a69cb294134d5bd3c16b4b15b56739c4fe3
Signed-off-by: Martijn Coenen <maco@google.com>