power: pm8921-bms: fix charging accuracy
The current driver assumes 100mA as the charge termination current.
Use the value configured from the board file.

Also the driver uses the battery voltage at the ADC to determine CC
to CV point. Although it works, it is more appropriate to calculate
battery terminal voltage and base CC to CV point on it.

CRs-Fixed: 420670
Change-Id: Ib046fd7149d71fc62c8f2df7a9747d6b54e95138
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>